Add fips mode test
