+++ tests/test_verify_build.py
@@ -5,7 +5,6 @@ Verify relenv builds.
 """
 import os
 import pathlib
-import platform
 import shutil
 import subprocess
 import sys
@@ -14,61 +13,13 @@ import textwrap
 import packaging
 import pytest
 
-from relenv.common import DATA_DIR, get_triplet, list_archived_builds, plat_from_triplet
-from relenv.create import create
+from relenv.common import DATA_DIR, get_triplet
 
-
-def get_build_version():
-    if "RELENV_PY_VERSION" in os.environ:
-        return os.environ["RELENV_PY_VERSION"]
-    builds = list(list_archived_builds())
-    versions = []
-    for version, arch, plat in builds:
-        sysplat = plat_from_triplet(plat)
-        if sysplat == sys.platform and arch == platform.machine().lower():
-            versions.append(version)
-    if versions:
-        return versions[0]
-
-
-@pytest.fixture(scope="module")
-def build_version():
-    version = get_build_version()
-    yield version
-
-
-@pytest.fixture(scope="module")
-def minor_version():
-    yield get_build_version().rsplit(".", 1)[0]
-
+from .conftest import get_build_version
 
 pytestmark = [
     pytest.mark.skipif(not get_build_version(), reason="Build archive does not exist"),
 ]
-
-
-@pytest.fixture
-def build(tmp_path, build_version):
-    create("test", tmp_path, version=build_version)
-    yield tmp_path / "test"
-
-
-@pytest.fixture
-def pipexec(build):
-    if sys.platform == "win32":
-        exc = build / "Scripts" / "pip3.exe"
-    else:
-        exc = build / "bin" / "pip3"
-    yield exc
-
-
-@pytest.fixture
-def pyexec(build):
-    if sys.platform == "win32":
-        exc = build / "Scripts" / "python.exe"
-    else:
-        exc = build / "bin" / "python3"
-    yield exc
 
 
 @pytest.mark.skip_unless_on_windows
